credit to https://github.com/Jasper9041/coconutphp
