+++ composer.json
@@ -14,7 +14,9 @@
         }
     ],
     "autoload": {
-        "psr-0": { "": "src/" }
+        "psr-4": {
+            "Coconut\\": "src/"
+        }
     },
     "require-dev": {
         "phpunit/phpunit": "4.3"
